add load emojis
